+++ src/rust/src/lazy/dsl.rs
@@ -1194,19 +1194,17 @@ impl RPolarsExpr {
         self.0.clone().arr().sum().into()
     }
 
-    // TODO: implement those in 0.38.0. They were wrongly included in the changelog of 0.37.0
-    // https://github.com/pola-rs/polars/issues/14355
-    // fn arr_std(&self, ddof: u8) -> Self {
-    //     self.0.clone().arr().std(ddof).into()
-    // }
+    fn arr_std(&self, ddof: u8) -> Self {
+        self.0.clone().arr().std(ddof).into()
+    }
 
-    // fn arr_var(&self, ddof: u8) -> Self {
-    //     self.0.clone().arr().var(ddof).into()
-    // }
+    fn arr_var(&self, ddof: u8) -> Self {
+        self.0.clone().arr().var(ddof).into()
+    }
 
-    // fn arr_median(&self) -> Self {
-    //     self.0.clone().arr().median().into()
-    // }
+    fn arr_median(&self) -> Self {
+        self.0.clone().arr().median().into()
+    }
 
     fn arr_unique(&self, maintain_order: bool) -> Self {
         if maintain_order {
@@ -1283,28 +1281,25 @@ impl RPolarsExpr {
             .into())
     }
 
-    // TODO: implement those in 0.38.0. They were wrongly included in the changelog of 0.37.0
-    // https://github.com/pola-rs/polars/issues/14355
-    // fn arr_to_struct(&self, fields: Robj, upper_bound: Robj) -> RResult<Self> {
-    //     let fields = robj_to!(Option, Robj, fields)?.map(|robj| {
-    //         let par_fn: ParRObj = robj.into();
-    //         let f: Arc<(dyn Fn(usize) -> SmartString<LazyCompact> + Send + Sync + 'static)> =
-    //             pl::Arc::new(move |idx: usize| {
-    //                 let thread_com = ThreadCom::from_global(&CONFIG);
-    //                 thread_com.send(RFnSignature::FnF64ToString(par_fn.clone(), idx as f64));
-    //                 let s = thread_com.recv().unwrap_string();
-    //                 let s: SmartString<LazyCompact> = s.into();
-    //                 s
-    //             });
-    //         f
-    //     });
-    //     Ok(RPolarsExpr(self.0.clone().arr().to_struct(fields)))
-    // }
+    fn arr_to_struct(&self, fields: Robj) -> RResult<Self> {
+        let fields = robj_to!(Option, Robj, fields)?.map(|robj| {
+            let par_fn: ParRObj = robj.into();
+            let f: Arc<(dyn Fn(usize) -> SmartString<LazyCompact> + Send + Sync + 'static)> =
+                pl::Arc::new(move |idx: usize| {
+                    let thread_com = ThreadCom::from_global(&CONFIG);
+                    thread_com.send(RFnSignature::FnF64ToString(par_fn.clone(), idx as f64));
+                    let s = thread_com.recv().unwrap_string();
+                    let s: SmartString<LazyCompact> = s.into();
+                    s
+                });
+            f
+        });
+        Ok(RPolarsExpr(self.0.clone().arr().to_struct(fields)))
+    }
 
-    // TODO: implement when bumping to rs-0.38.0
-    // fn arr_shift(&self, n: Robj) -> RResult<Self> {
-    //     Ok(self.0.clone().arr().shift(robj_to!(PLExprCol, n)?).into())
-    // }
+    fn arr_shift(&self, n: Robj) -> RResult<Self> {
+        Ok(self.0.clone().arr().shift(robj_to!(PLExprCol, n)?).into())
+    }
 
     // datetime methods